// src/monitoring/enums/alert-type.enum.ts
export enum AlertType {
  HIGH_CANCELLATION_RATE = 'high_cancellation_rate',
  LOW_RIDER_AVAILABILITY = 'low_rider_availability',
  HIGH_ERROR_RATE = 'high_error_rate',
  SLOW_RESPONSE_TIME = 'slow_response_time',
  SYSTEM_DOWN = 'system_down',
  PAYMENT_GATEWAY_ISSUE = 'payment_gateway_issue',
  SMS_SERVICE_ISSUE = 'sms_service_issue',
  DATABASE_CONNECTION_ISSUE = 'database_connection_issue',
  HIGH_ORDER_VOLUME = 'high_order_volume',
  STORE_INACTIVE = 'store_inactive',
}